Add files via upload
All notebooks concerning the construction of the Index. First (Data_Handling)includes data gathering, cleaning, handling of data bases and management of the index. Second (price) is the introduction of prices and basic capitalization method for the listed companies in the index. Third (Free_float_adj) introduces the free float mechanism to the index. "funcs_tfm" are the utilities used in the making of the index that are not included in the notebooks.